Remove AstVisitor trait
