+++ Cargo.toml
@@ -1,3 +1,4 @@
 [workspace]
 members = ["clar2wasm", "tests"]
 exclude = ["stacks-blockchain"]
+resolver = "2"
